renaming and clean code up
