+++ src/fetchLocations.js
@@ -1,4 +1,4 @@
-export default async function getCoordinates({ url, parser }) {
+export default async function fetchLocations({ url, parser }) {
   const response = await fetch(url);
   const data = await response.json();
