+++ src/index.js
@@ -3,7 +3,7 @@ import ReactDOM from 'react-dom';
 import MuiThemeProvider from 'material-ui/styles/MuiThemeProvider'; // Theme
 import injectTapEventPlugin from 'react-tap-event-plugin'; // Click handler
 import './styles/index.css';
-import locationHistoryParser from './locationHistoryParser';
+import locationHistoryParser from './parseLocations';
 import {
   GM_ZOOM,
   GM_CENTER_LAT,
@@ -14,7 +14,7 @@ import {
   GM_VISUALIZATION,
   JSON_FILE_URL
 } from './helpers/constants';
-import getCoordinates from './locationHistoryController';
+import fetchLocations from './fetchLocations';
 import LocationHistoryMap from './components/LocationHistoryMap';
 import Header from './components/Header';
 
@@ -29,7 +29,7 @@ export default function init() {
     }
   };
 
-  getCoordinates({
+  fetchLocations({
     url: JSON_FILE_URL,
     parser: locationHistoryParser
   }).then(coordinates => {
